WIP: Updates test configuration
Tests that run against floated versions are allowed to fail because it is not a regression in the changes.
Thus it makes little difference if the changes are allowed to be submitted and then the problem is fixed or the other way around.

Code quality tests run in py37 making that the primary version of the project.