Customizes docs
Includes readme on index page to make the docs not completely useless fast.

Makes sphinx-apidoc run on readthedocs by adding an extension that runs it
automatically whenever the docs are built. Not having to manually include
modules etc in the docs should make it easier to maintain even if it will not
be as neat.

Makes readthedocs use python 3 because that is what this project targets, and
pip install the package so that autodoc works.