Adds iterutils.imerge
... to facilitate mergin sorted iterators that should not be exhausted.

The use case I have in mind is doing a bf search of git commits.

Adds `T` to the list of good names because it is commonly used for type
variables.

Adds `v` to the list of good names because it is convenient as a placeholder for
a generic value in a small scope.

Removes `isort` tests because they disagree with yapf.