Improves coverage testing
Uses only the installed package when collecting coverage information to make the report
more reliably reflect the coverage on code that is shipped.

Checks branch coverage as well to get a more complete view of what code paths have not
been tested.

Lists the longest running tests to make it easier improving the test suite and lowering
the expected time to failure and the expected time to success.

Explicitly marks coverage collecting envs as dependencies for the reporting envs to
avoid race conditions.

f