Drops py35 support
Because I no longer need it.

Replaces old style type hints because I can and to get rid of pylint warnings.

Upgrades pylint to get rid of false positive under py38.

Upgrades pip and pip-tools to be able to upgrade pylint.

Upgrades other packages for good measure and in hope that it will help CI get unstuck.

Tests on py36 instead of pypy3 on windows because I cannot be bothered getting it to
work. The error message is not very good, but implies sprig cannot be packaged:

> ERROR: FAIL could not package project ...