Enables type checking of external packages
I had previously forgotten that dependencies need to be installed in order to be found
during type checking. I had also been to lazy to find out why more_itertools could not
be found and just ignored it.

Also resolves type problem that mypy complains about now that it is able to check more
lines.