Adds unmanaged merger
Adds the `TimeoutMerger` class that automatically registers and unregisters senders,
hence the use of the word _unmanaged_. Any use case that leverages the managed merger
would (as opposed to one of the static implementations) will register and unregister
senders. The implemented policy for doing so is the simplest one I can think of but
serves to demonstrate the classes. It is also one that I would like to come up with
a more satisfying implementation for since I don't like the current one that runs in
linear time or worse.

This commit also make the following changes that would perhaps best be factored out to
their own commits but I am not in the mood:

Starts phasing out the term bucket as it is typically used for unsorted collections
whereas here we expect the opposite.

Rewrites the merger without key functions to make the code easier to follow and,
theoretically, more performant. It is easy to implement the old interface on top of the
new one at little cost to readability or (theoretically) performance. Indeed this has
been done in order to avoid changing existing tests.

Improves validation of arguments to alert users to bugs often, early, and close to the
cause.

Reframe the implementation in terms of clients sending messages in order and a server
receiving them. This metaphor is more concrete and easier to visualize. The generality
of the algorithm is maintained using type hints.