Defines interval intersection for edge cases
This was prompted by my rust implementation not disagreeing with my python
implementation because I had assumed the now cemented behavior was the one already
implemented.

As the comments explain, when faced with the choice of whether to include things or
exclude them, if both options are cheap to implement and cheap to run then I have
chosen the behavior that is cheaper to adjust by a consumer.