Upgrades dependencies
... to avoid using vulnerable versions and get a set of constraints that work together
to pass CI.

Formats files because newer version of black requires it.

Re-raise from to make pylint happy.

Replace m2r because it causes sphinx to crash and appears abandoned.

Constrain pip versions because newer versions dislike hashes in constraints files and
I am not yet sure if this is a temporary issue.

Constrain virtualenv version because newer versions install a newer version of pip.

Note that the pip situation is a bit inconsistent; we bootstrap the virtual machine to
20.0.2, upgrade to 20.3.4 from constraints file then tox (via virtualenv) uses 20.1.1.